refactor: fix: moved predefined models to dinjo/predefined/epidemiology
