+++ dinjo/predefined/epidemiology/__init__.py
@@ -1,0 +1,11 @@
+from ._seir_model import ModelSEIR
+from ._seirv_model import ModelSEIRV
+from ._seirv_fixed import ModelSimpleSEIRV
+from ._sir_model import ModelSIR
+
+__all__ = [
+    'ModelSEIR',
+    'ModelSEIRV',
+    'ModelSimpleSEIRV',
+    'ModelSIR',
+]
