+++ dinjo/predefined/epidemiology/_seirv_model.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from ..model import ModelIVP
+from ...model import ModelIVP
 
 
 class ModelSEIRV(ModelIVP):
